requirements.txt
playwright==1.44.0
httpx==0.27.0
pydantic==2.7.1
supabase==2.4.6
python-dotenv==1.0.1
pytest==8.2.0
pytest-asyncio==0.23.6

fastapi==0.111.0
uvicorn==0.30.1
